Split loaders.py into three modules
